Interfering task

+++ src/inc/prefetch.h
@@ -1,7 +1,7 @@
 #ifndef __PREFETCH_H__
 #define __PREFETCH_H__
 
-void clear_L2_cache(unsigned long long address, unsigned long long size);
-void prefetch_data(unsigned long long address, unsigned long long size);
+void clear_L2_cache(uint64_t address, uint64_t size);
+void prefetch_data(uint64_t address, uint64_t size);
 
 #endif
